Resolve UX review comments.
