+++ packages/terra-framework-docs/src/terra-dev-site/doc/worklist-data-grid/About.1.doc.mdx
@@ -3,8 +3,6 @@ import { Badge } from 'terra-grid/package.json?dev-site-package';
 import WorklistDataGridPropsTable from 'terra-worklist-data-grid/src/WorklistDataGrid?dev-site-props-table';
 
 import DefaultWorklistDataGrid from './example/DefaultWorklistDataGrid?dev-site-example';
-import WorklistDataGridMaskedCell from './example/WorklistDataGridMaskedCell?dev-site-example';
-import WorklistDataGridWithSelections from './example/WorklistDataGridWithSelections?dev-site-example';
 
 <Badge />
 
@@ -37,10 +35,6 @@ This component requires the following peer dependencies be installed in your app
 
 ## Examples
 <DefaultWorklistDataGrid title= 'Default Worklist Data Grid' />
-
-<WorklistDataGridMaskedCell title= 'Worklist Data Grid with Masked Cells' />
-
-<WorklistDataGridWithSelections title= 'Worklist Data Grid with Selections' />
 
 ## Grid props table
 <WorklistDataGridPropsTable />
