--- packages/terra-framework-docs/src/terra-dev-site/doc/worklist-data-grid/example/WorklistDataGridMaskedCell.jsx
@@ -1,72 +1,0 @@
-import React from 'react';
-import WorklistDataGrid from 'terra-worklist-data-grid';
-
-const gridDataJSON = {
-  cols: [
-    { id: 'Column-0', displayName: 'Patient' },
-    { id: 'Column-1', displayName: 'Location' },
-    { id: 'Column-2', displayName: 'Illness Severity' },
-    { id: 'Column-3', displayName: 'Visit' },
-    { id: 'Column-4', displayName: 'Allergy' },
-    { id: 'Column-5', displayName: 'Primary Contact' },
-    { id: 'Column-6', displayName: 'Generic Order Counts' },
-    { id: 'Column-7', displayName: 'Patient Age' },
-    { id: 'Column-8', displayName: 'Medication History' },
-    { id: 'Column-9', displayName: 'My Relationship' },
-    { id: 'Column-10', displayName: 'Forms' },
-  ],
-  rows: [
-    {
-      id: '1',
-      cells: [
-        { content: 'Fleck, Arthur' },
-        { content: '1007-MTN' },
-        { content: 'Unstable' },
-        { content: 'Inpatient, 2 months' },
-        { content: '' },
-        { content: 'Quinzell, Harleen' },
-        { content: '' },
-        { isMasked: true },
-        { isMasked: true },
-        { content: 'Admitting Physician' },
-        { content: '' },
-      ],
-    },
-    {
-      id: '2',
-      cells: [
-        { content: 'Wayne, Bruce' },
-        { content: '1007-MTN-DR' },
-        { content: 'Stable' },
-        { content: 'Outpatient, 2 days' },
-        { content: 'Phytochemicals' },
-        { content: 'Grayson, Richard' },
-        { content: '' },
-        { content: '' },
-        { isMasked: true },
-        { content: 'Admitting Physician' },
-        { content: '' },
-      ],
-    },
-  ],
-};
-
-const WorklistDataGridMaskedCell = () => {
-  const { cols, rows } = gridDataJSON;
-
-  return (
-    <React.Fragment>
-      <WorklistDataGrid
-        id="terra-worklist-data-grid-masked-cell"
-        columns={cols}
-        rows={rows}
-        rowHeight="50px"
-        columnWidth="50px"
-        columnHeaderHeight="50px"
-        ariaLabel="Worklist Data Grid"
-      />
-    </React.Fragment>
-  );
-};
-
-export default WorklistDataGridMaskedCell;
