+++ packages/terra-worklist-data-grid/src/WorklistDataGrid.jsx
@@ -172,7 +172,7 @@ function WorklistDataGrid(props) {
       <WorklistCellTag
         key={cellColumnIndex}
         {...tabIndex}
-        className={cx('worklist-data-grid-cell', { masked: cell.isMasked, selectable: (!cell.isMasked && cell.isSelectable) })}
+        className={cx('worklist-data-grid-cell', { masked: cell.isMasked, selectable: !(cell.isMasked || cell.isSelectable === false) })}
         aria-label={cell.isMasked ? intl.formatMessage({ id: 'Terra.worklistDataGrid.maskedCell' }) : undefined}
       >
         <div className={cx('cell-content')}>{cell.content}</div>
@@ -185,7 +185,7 @@ function WorklistDataGrid(props) {
     const height = props.columnHeaderHeight;
     return (
       /* eslint-disable react/forbid-dom-props */
-      <th key={column.id} className={cx('worklist-data-grid-column-header', { selectable: column.isSelectable })} tabIndex="-1" style={{ width, height }}>{column.displayName}</th>
+      <th key={column.id} className={cx('worklist-data-grid-column-header', { selectable: !(column.isSelectable === false) })} tabIndex="-1" style={{ width, height }}>{column.displayName}</th>
     );
   };
