+++ packages/terra-worklist-data-grid/tests/jest/WorklistDataGrid.test.jsx
@@ -3,7 +3,41 @@ import React from 'react';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
 import { shallowWithIntl } from 'terra-enzyme-intl';
 import WorklistDataGrid from '../../src/WorklistDataGrid';
-import dataFile from './testData.json';
+
+// Source data for tests
+const dataFile = {
+  cols: [
+    { id: 'Column-0', displayName: ' Vitals' },
+    { id: 'Column-1', displayName: 'March 16' },
+    { id: 'Column-2', displayName: 'March 17', isSelectable: false },
+  ],
+  rows: [
+    {
+      id: '1',
+      cells: [
+        { content: 'Heart Rate Monitored (bpm)' },
+        { content: '68' },
+        { content: '66', isMasked: true },
+      ],
+    },
+    {
+      id: '2',
+      cells: [
+        { content: 'Temperature Oral (degC)' },
+        { content: '36.7' },
+        { content: '36.9', isMasked: true },
+      ],
+    },
+    {
+      id: '3',
+      cells: [
+        { content: 'Cardiac Index (L/min/m2)' },
+        { content: '2.25' },
+        { content: '2.28', isMasked: true },
+      ],
+    },
+  ],
+};
 
 describe('WorklistDataGrid', () => {
   it('renders a grid with 3 columns and 3 data rows and 9 cells', () => {
@@ -16,10 +50,10 @@ describe('WorklistDataGrid', () => {
     ).dive();
 
     const columnHeaders = wrapper.find('.worklist-data-grid-column-header');
-    const rowHeaders = wrapper.find('.worklist-data-grid-cell');
+    const dataGridCells = wrapper.find('.worklist-data-grid-cell');
 
     expect(columnHeaders).toHaveLength(3);
-    expect(rowHeaders).toHaveLength(9);
+    expect(dataGridCells).toHaveLength(9);
     expect(wrapper).toMatchSnapshot();
   });
